Record the open PRs and unassigned MessageIDs
In order to associate original commits with commits in git.git, we will
need to keep track of the sent patch mails for which we do not yet know
of any such corresponding commit.

While at it, also keep track of the open Pull Requests. This will be
needed to be able to close Pull Requests when their corresponding
patches were merged to `master`.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>