Add the main work-horse for the automated "CI jobs"
While the normal purpose of "Continuous Integration" jobs is to monitor
branches and perform tests whenever they are updated (or Pull Requests
are opened), to verify that no regressions were introduced, we can also
use the same technology to monitor e.g. the `mail-to-commit` ref and
trigger actions based on what we find.

For example, we want to know when our patches have been applied and
integrated into, say, `pu`, by monitoring and parsing both the
`refs/notes/mail-to-commit` ref as well as the `refs/heads/pu` branch.

This commit adds the CIHelper class and a script to call its methods,
but at the moment it does not do a whole lot of anything (except that
CIHelper already has the fun part of determining which commit, if any,
merged a branch into another branch). For good measure, we also add some
tests.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>