CIHelper: detect into which upstream branch a PR was merged, if any
This information requires quite a bit to be detected before that: which
original commits were already integrated into git.git and what is the
new SHA-1 of the upstream commits.

While at it, we can also detect which name Junio Hamano (the git.git
maintainer) chose for the patch series, and we can also update the
GitHub PR's labels to reflect which upstream branch(es) integrated the
patch series already.

And we can also close the PR once the patch series hit `master` or
`maint`.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>