Add a script to generate the mail->commit mapping
This is a companion to `lookup-commit.sh --notes update`: after that
command updated refs/notes/commit-to-mail, this here script will update
refs/notes/mail-to-commit to provide a reverse mapping: for every
Message-ID, one can now find the active git.git commit thusly:

	hash="$(echo "$MESSAGE_ID" | git hash-object --stdin)"
	git notes --ref=ref/notes/mail-to-commit show $hash

If the command turns up empty-handed, there is no active git.git commit
for that mail.

Note: in this context, a git.git commit is "active" as long as it is
reachable from `pu`. Whenever `pu` is updated and formerly-reachable
commits become unreachable, their mapping will be removed from the notes
(although you can still see them in the history of that ref).

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>