ci-helper: offer to update commit mappings
For all open PRs, we want to see whether the commits have counterparts
in git.git. This commit adds a one-stop task to do exactly that.

It does rely on the GitGitGadget options being kept up to date with the
open PRs and the corresponding Message-IDs, and since this patch is
developed on a topic branch, the GitGitGadget version running at the
time of writing does not do that. For that reason, this task currently
requires to be run after the "update-open-prs" task has been run.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>