Clarify `From:` and add `Sender:`  Email fields
For GGG, record the user who did the pull request on the "From:"
header of cover letter and patches in the series.

Only add in-body From: as necessary to record the commit author name
when it is different from the user who created the pull request.

Use "Sender:" to record the fact that the message is sent from
gitgitgadget@gmail.com.

Signed-off-by: Chris. Webster <chris@webstech.net>
Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>