CIHelper: add a helper function to support the upcoming GitHub Action
The plan is to turn GitGitGadget into a set of GitHub Actions that can
be run conveniently from GitHub workflows.

To that end, we will add subdirectories (e.g. `init/`, `handle-prs/`,
etc) that reflect the individual functionality that is currently called
in Azure Pipelines via subcommands of the `misc-helper` script.

Those subdirectories will contain the bare minimum required to implement
a Javascript Action that calls into the CIHelper class: an `action.yml`
file to declare the inputs/outputs, and a minimal `index.js` that
consumes the inputs and passes them to the CIHelper.

The CIHelper class will be used by the Action via as a single,
dependency-free `dist/index.js`, and it will be the only dependency of
the Action's `index.js`.

In particular, we must not let the Action's minimal `index.js` depend on
`@actions/core` because it simply won't be there.

So let's add it as a dependency of the CIHelper class and provide a
function to allow calling the imported `@actions/core` functions from
the Actions' `index.js`.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>