Ensure that `res/WELCOME.md` is found even after compiling with `ncc`
According to https://github.com/vercel/ncc/issues/829, `ncc` has some
magic that includes the resource in the bundled package if the path is
specified using `path.join(__dirname, ...)`.

However, we're already using ESModule, therefore we need to use the
`import.meta.url` syntax (`import.meta.dirname` unfortunately does not
seem to work in my hands).

Having said that, the promise was that `ncc` would include this asset in
the output, but this is not true for me.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>