GitGitGadget: be intentional about the push URL in addition to the token
I just changed the way the push token is handed down from CIHelper to
GitGitGadget instead of relying on the Git config (or environment) to
specify it implicitly. Let's do the same for the URL of the repository
to which to push the Git notes and patch series iterations' tags.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>