CIHelper.setupGitHubAction(): also process the SMTP inputs, if any
The upcoming `handle-pr-comment` Action wants to send emails, and for
that reason needs to accept SMTP information. This will be passed into
the Action via the `smtp-options` input, which is a JSON-formatted
triplet of user/host/pass info.

Since most of GitGitGadget's Actions do not want to send emails, it is
not an error if there are no `smtp-options` when `setupGitHubAction()`
is called.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>