ci: automatically bundle & push the GitHub Actions to the v1 branch
We converted GitGitGadget into a bunch of GitHub Actions. The core logic
of these is implemented in Typescript, which means that the code needs
to be transpiled and bundled before it can be used as GitHub Action.

This new workflow does exactly that. It maintains a new branch, called
`v1`, and incremental tags of the `v1` tip commits, that are updated
whenever the `main` branch advances. The updates consist of merging in
`main`'s commit history, but the actual tree contents reflect the
transpiled output, ready to be called as a GitHub Action from GitHub
workflows.

To reduce bandwidth when using the Action, all source code is removed
from these tip commits, except for the few resources (`WELCOME.md` and
two shell scripts) that are moved into the location where they need to
be in order to be picked up by the GitHub Actions whenever needed.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>