amend! Add a workflow to push the GitHub Actions
ci: automatically push the bundled GitHub Actions to the v1 branch

We converted GitGitGadget into a bunch of GitHub Actions. The core logic
of these is implemented in Typescript, which means that the code needs
to be transpiled and bundled before it can be used as GitHub Action.

This new workflow does exactly that. It maintains a new branch, called
`v1`, and incremental tags, that are updated whenever the `main` branch
advances. The updates consist of merge commits that merge in `main`'s
commit history, but whose trees reflect the transpiled output (and all
source code removed, to reduce bandwidth).

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>