CIHelper: optionally clone the Git mailing list mirror
A couple of GitGitGadget's common operations require a local clone of
the Git mailing list mirror.

For obvious reasons, the idea here is to initialize a partial, shallow
clone, so that missing objects (if any) will be retrieved on demand, and
no time is wasted on a full clone every time GitGitGadget wants to look
at mails.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>