CIHelper: offer to pre-populate the upstream branches
GitGitGadget's code has auto-magic logic to fetch those, but even in a
partial clone that would result in a ~100MB fetch. Maybe that's okay,
and I worry too much about bandwidth? Inside hosted Actions runners,
download speeds of ~30MB/sec are not unheard of, after all... But then,
I do think that too few people are careful to waste less resources than
is ethical, so I want to be mindful.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>