lookup-commit.sh: stop pretending that git/git is gitster/git
That script wants to ensure that the remote called `gitster` is
registered, but when it is not, it registers it with the URL
https://github.com/git/git. That is misleading.

All the script wanted from the gitster/git repository was the `seen`
branch, anyway, and the git/git repository does have that.

So let's just refer to it as `upstream/seen`, as is done elsewhere in
GitGitGadget's code base.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>