CIHelper: offer to update the `mail-to-commit` and `commit-to-mail` notes
This ports the Azure Pipeline "Update GitGitGadget's commit to mail
notes" (https://dev.azure.com/gitgitgadget/git/_build?definitionId=9)
logic to a method of the `CIHelper` class that will be used in a GitHub
Action so that the Azure Pipeline can be retired.

This was the only Azure Pipeline of GitGitGadget which did not,
actually, use the `misc-helper.ts` script, but instead was opaquely
implemented as a lengthy shell scriptlet inside the Pipeline definition
that called two of the shell scripts in the `gitgitgadget/gitgitgadget`
repository: `lookup-commit.sh` and `update-mail-to-commit-notes.sh`.

Since the scripts that are called by the new method expect a persisted,
non-partial clone of the Git mailing list, we need to play a couple of
games here to make it work in a GitHub workflow (that runs on ephemeral
runners where the repository has to be cloned afresh in every run).

Also: These shell scripts, by virtue of needing to be interpreted by a
shell interpreter outside of node.js, need to be copied into the
`dist/script/` subdirectory before publishing the GitHub Action, so that
they can be found and interpreted as expected.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>