CIHelper: offer to update the `mail-to-commit` and `commit-to-mail` notes
This ports the Azure Pipeline "Update GitGitGadget's commit to mail
notes" (https://dev.azure.com/gitgitgadget/git/_build?definitionId=9)
logic to a method of the `CIHelper` class that will be used in a GitHub
Action so that the Azure Pipeline can be retired.

Since the scripts that are called by the new method expect a persisted,
non-partial clone of the Git mailing list, we need to play a couple of
games here to make it work in a GitHub workflow (that runs on ephemeral
runners where the repository has to be cloned afresh in every run).

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>