Avoid potentially expensive fetching of tags
Right now, there is no problem: GitGitGadget uses a couple of Azure
Pipelines that all run on the same, single runner inside a dedicated
pool, and this runner has a persistent clone of the `gitgitgadget/git`
repository.

However, my current plan is to migrate to GitHub Actions, away from
that self-hosted runner, which means that there is no persistent clone
to work with. Therefore, a local Git repository has to be initialized in
every workflow run, and to make that efficient (and fast!), only the
minimal set of refs is fetched. That idea works only when preventing
Git from _implicitly_ fetching all tags, so let's do that.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>