ProjectOptions: `baseCommit` is actually mandatory
Back in the days when I tried to maintain a Command-Line Interface in
GitGitGadget to replace my `mail-patch-series.sh` script (still
available at https://github.com/dscho/mail-patch-series, it was the
predecessor of GitGitGadget that worked on local repositories instead of
Pull Requests), there was not necessarily a `baseCommit` to work with:
local branches can have an upstream branch (configured via `git branch
--set-upstream-to=<remote>/<branch>`).

But I removed support for this way to call GitGitGadget in 6d397b0f
(Remove support for the CLI mode, 2023-12-31), so now there always is a
`baseCommit`.

Let's make it mandatory, so that it can eventually take on the role of
the `upstreamBranch` (which we can then remove).

Since optional parameters must come after non-optional ones, this
unfortunately requires a reordering of the (positional) parameters of
the `ProjectOptions.get()` function.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>