IConfig: remove no longer needed `project.branch`
Ever since 6d397b0f (Remove support for the CLI mode, 2023-12-31), there
is no way the `ProjectOptions` are constructed without a `baseCommit`.
Therefore it is always provided.

And since `baseCommit` is always provided, the `ProjectOptions.get()`
function no longer needs `upstreamBranch`: It only used it in a "is this
rebased?" check that apparently hasn't been called anymore, ever.

So that leaves just one last user of that `upstreamBranch` attribute,
but that's a test that wants to mocj a `prMeta` from a `ProjectOptions`
instance, something that only the CLI mode used to do, but no remaining
production code. That test can easily replace that
`options2.upstreamBranch` with a sensible, hard-coded string.

And since that's gone, we can also remove the `project.branch` attribute
because nothing uses it anymore.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>