Set `project` in the `defaultConfig`
We're about to make the config required, and not implicit, therefore it
makes sense to start with the project options for `gitgitgadget/git`
itself.

So let's set `project` explicitly in `defaultConfig` and remove the
implicit configuration of the `gitgitgadget/git` project from
`ProjectOptions.get()`.

There is a slight quirk here, as the Git GUI special-casing is a bit
magic: based on the presence of a specific file at the top-level of
`HEAD`'s root tree, it is determined that this is a Git GUI PR rather
than a git/git PR.

Now, we _could_ introduce some new attribute of the `projectInfo` type
that would allow the same kind of trick for other projects. As per
https://github.com/gitgitgadget/gitgitgadget/pull/1988#issuecomment-3263190977
though, no other project using GitGitGadget may ever need that, so we'll
still hard-code it for `gitgitgadget/git` (and `git/git`) specifically.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>