IPatchSeriesMetadata: the PR URL is not actually optional
It _used_ to be optional when there was also a way to run GitGitGadget
in a local repository on a branch that had an upstream branch
configured, in which case there was no PR to work with.

But after I dropped the CLI in 6d397b0f (Remove support for the CLI
mode, 2023-12-31), GitGitGadget really only works with PRs.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>