Retire `misc-helper`
That script had been GitGitGadget's work horse back in the days when the
logic was performed in Azure Pipelines.

However, in the meantime, GitGitGadget migrated to GitHub workflows,
where the main work horse is the `CIHelper` class instead. See
https://github.com/gitgitgadget/gitgitgadget-github-app/pull/4 for the
full details.

That migration has now been completed, and in production for the past
two months. It's therefore safe to say that we do not need the
`misc-helper` script anymore.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>