check-run: persist parameters in an Action state
... and reuse it when called a second time. That way, one can update,
say, only the text, without having to re-specify all the details
_again_.

For technical details, the token and the config _do_ need to be
re-specified again (the token because of security concerns, the config
so that the appropriate token can be used just in case that pr-repo and
upstream-repo can be handled by the same workflow).

Note: Instead of using `core.saveState()`, the state is explicitly
persisted via the environment variable `STATE_check-run` (yes, it _is_ a
funny name). The reasons for that are:

1. Subsequent invocations of the Action would _not_ get the state, only
   the `post` Action would.
2. The `post` Action would only receive the state saved by the _first_
   Action invocation, subsequent `saveState()` calls would be ignored!

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>